Implemented Leaderboard cache

+++ Game.Ranking.Model/Model/ReplicableObject.cs
@@ -6,7 +6,7 @@ namespace Game.Ranking.Model
     public abstract class ReplicableObject
     {
         [Key]
-        public int Key { get; set; }
+        public long Key { get; set; }
 
         public long InsertedTimestamp { get; private set; }
         public long ReplicatedTimestamp { get; private set; }
